refactor(cdk-explorer): extract assembly read-lock factory into core
Move the AssemblyLock type and the fromAssemblyDirectory().produce()-backed
acquirer out of lib/lsp into lib/core/assembly-lock.ts, so both the LSP and the
web server can build the read lock from one shared factory. Pure refactor: LSP
behavior is unchanged and server.ts re-exports AssemblyLock for existing
importers.
